Specify Hadoop jobs using Scala functions.
Implement framework allowing a Hadoop Job to be created and run by
specifying two Scala functions - 'map' and 'reduce'. The functions use
"normal" types (i.e. not Hadoop Writable types) but have type-class
constraints on them for serialization (HadoopWritable) and Ordering, in
the case of keys.

Because Hadoop jobs require Writable and WritableComparable subclasses
for values and key types respectively, the Scala types used in the
'map' and 'reduced' functions must be converted. This is done by using
the type-class instance objects to build specialised Writable and
WritableComparable classes at run-time (ScoobiWritable and
ScoobiWritableComparable). The Javassist library is employed to do this.

Since new classes are generated for each Hadoop job, they must be
included in the job's JAR. A class, JarBuilder, is used to manage the
construction of JAR, purpose-built for each Hadoop job, containing
generated classes.

Generic Mapper and Reducer objects work with the generated key and value
types. These objects are serialized and pushed out to the cluster using
Hadoop's distributed cache feature (DistributedObject).