Handle primivite types in runtime class building.
In compiling some of the class-type models, it's possible that Scala
makes use of Java primitive types. As a consequence, when these models
are accessed within a generated ScoobiWritable* class (i.e. a Java class
built at runtime), the possiblity of primitive types must be handled.

Modify the code generation to catch the case of primitive types and
handle it by wrapping and unwrapping their associated "wrapper" types
(e.g. Integer for int, etc).