Implement I/O at the language level.
Add APIs that to create distributed lists by reading in text files using
a mixture of type-classes and pattern-matching. Also add APIs that take
one or more distributed lists and persists them as text files. The act
of persisting a distributed list becomes the catalyst for excuting the
computation graph behind it.