Create Hadoop jobs from MSCRs.
Define an MSCR (Map-Shuffle-Combine-Reduce) data structure (AST and
MSCR) as an intermediate representation of a distributed-list
computation graph (Smart). From each MSCR, construct a MapReduceJob that
includes the functionality specified by the MSRC, which in turn
refrences nodes within a distributed-list AST (AST).

Implement an excution engine (Executor) that is responsible for taking a
collection of MSCRs then generating and running MapReduce jobs for each.
The collection of MSCRs actually form a graph that is traversed by the
execution engine, which is also responsible for ensuring that each MSCR
is only executed once and that intermediate outputs are garbage
collected.

The construction of MapReduce jobs from MSCRs requires the runtime
generation of a series of classes (TaggedKey, TaggedValue,
TaggedPartitioner) that implement an aggregated K2-V2 types for each
intermedate types used by all mappers within an MSCR. Type-class
instances (Manifest, HadoopWritable, Ordering) are used to assist in the
generation of these run-time classes.