Add JobConf to Scoobi object.
Have the Scoobi object create a JobConf value that can be used
throughout the code base to discover properties such as the filesystem.
Use this JobConf as the starting point for creating the JobConf used for
each MapReduce run.

Looking forward, it may make sense for the Scoobi object to direclty
subclass Configuration. In the long run this may be more useful.