Store outputs as text rather than sequence files.
In order to store outputs as text, allow output nodes (e.g. combiners,
reducers) in a given MSCR output channel to store to more than one
output file. Using this approach, a node that is both a global output
as well as an input to another MSCR will write out to both a text file
and a sequence file.

This implementation results in a renaming of DConnector to DataStore,
and InputLike/OutputLike to DataSource/DataSink. Of particular note, the
OutputStore class no longer mixes in the DataSource trait - it is only a
sink.