Don't create closure when fusing FlatMaps.
If a closure is created when fusing the mapping functions of two
successive Smart.FlatMap nodes, the resulting mapping function will be a
Function1 type object with a reference to the outer FlatMap node. This
same function object will be referenced in an AST.Mapper or
AST.GbkMapper node which will mean everything in the original
Smart.FlatMap node will be part of its object graph. The problem with
this is that when mapping functions are serialised, their entire object
graph is serialised too!

Rather than creating a closure when fusing the FlatMap funnctions,
simply construction a new function (Function1) that is a "composition"
of the two function objects from each FlatMap.