Implement support for FlatMap only graphs.
Implement support for the case where an output is simply a FlatMap
preceeded by a Load. If the Load is not an input shared by other parts
of the graph it will likely be missed by the standard related-GBK
strategy for forming MSCRs. In this case, find all outputs that have not
been accounted for and place them into a "simple" MSCR.