Set number of reducers with a simple heuristic:
Base the amount of parallelism required in the reduce phase on the size
of the data output. Further, estimate the size of output data to be the
size of the input data to the MapReduce job. Then, set the number of
reduce tasks to the number of 1GB data chunks in the estimated output.