Support generic Hadoop command line options.
Support command line options such as -fs and -jt by introducing a
"control structure" called 'withHadoopArgs'. It takes two parameter
lists: the first are the command line options, and the second is a
function which is the user's code. Using 'withHadoopArgs' will set the
JobConf based first on $HADOOP_CONF and then by any overrides provided
on the command line.