Add cast in TaggedValue's call to fromWire.
In a TaggedValue implementation of readFields there are calls to
fromWire. In the case of the Serializable implementation of WireFormat
that fromWire will return T where T is a subclass of Serializable. In
this case a cast is required from T to the actual type of the value else
a ClassVerify RTE will occur.