Implement 'materialize' method for DLists.
Add a 'materialize' method on DLists that returns a DObject of type
Iterable. That is, it is a mechansim to bring the values of a DList out
of a Scoobi job/computation and into the client world.

A DObject is a wrapper type for "scalar" result computations within a
Scoobi job. Its API is somewhat brief and is likely to change.

To specify that a DObject is to be included as part of a Scoobi job's
computations, provide a 'Job' class, to which persisted DLists and
DObjects can be added, after which it can be run (forcing the
computation).