Assign reducer tasks to output specific channels.
Each reducer tasks must only be responsible for key-values associated
with a single output channel. This is to ensure that the results of each
output channel are written to isolated sets of output files. To ensure
this make two changes:

1. Modify the calculation of the job's reducer task count to be at a
minimum the number of output channels in a given MSCR;

2. Specify the same Partitioner class for each job (rather than
generating one) that partitions solely by the key-value's tag.