Make finding related nodes less greedy.
In creating MSCRs, nodes that are determined to be "related" are found
and a group of such related nodes forms the creation of an MSCR. The
previous implementation was too greedy in how it deemed nodes to be
related resulting in nodes being inclded in MSCRs together when they
should have been separated. In particular, two GroupByKey nodes could be
considered "related" even when there was a dependency between them.

Modify the "related"-ness algorithm to be less greedy and in effect be
correct for the formation of MSCRs. The new algorithm will implement
three consequtive passes that first check for relatedness by ParallelDo
nodes, then by ParallelDo node inputs, and finally pruning out PrallelDo
nodes that are unrelated due to GroupByKey dependencies.