Specify Java class for primitive (AnyVal) outputs.
If the type of a DList being persisted using toTextFile is an AnyVal,
Manifest[A] will resolve to the primivite Java type. For example,
Manifest[Int] will resolve to 'int'. Of course, 'int' is not a class so
in this case, and for all other AnyVal types, use the Java class
equivalent. That is, rather than Manifest[Int], use java.lang.Integer.