Use XStream for serialization.
Scoobi serializes objects for transportation to Hadoop task trackers via
the distributed cache. An issue with using Java serialization is that
the all classes in the object graph must be Serializable. This is a
problem when using 3rd-party libraries where classes do not extend the
Serializable interface.

In contrast, the XStream serialization library does not require classes
to be Serializable in order to serialize them. Modify the 'pushObject'
and 'pullObject' methods in 'DistCache' to serialize/deserialize objects
to/from and XML string.