Add ScoobiWritable to Job JAR for inputs.
DLists that span between multiple MSCRs are persisted as Sequence Files
of ScoobiWritable objects. The class definitions for these objects are
generated dynamically and must be added to a MapReduce job's "user" JAR.
Fix a bug where the class files for ScoobiWritable's are only added to
the JAR when the DList is an output of a MapReduce job and not when it
is an input.