Add 'fuse' and 'grouping' barriers to ParallelDo.
Add flags to the intermediate ParallelDo node type that specify 'fusing'
and 'grouping' barriers.

A ParllelDo node marked with a 'fuse' barrier indicates that a successive
ParallelDo node can not be fused with it. By default, mark all output
ParallelDo nodes with 'fuse barriers.

A ParallelDo node marked with a 'grouping' barrier indicates that it and
all previous nodes, up to a GroupByKey node, must not be split accross
MapReduce jobs. Expose this barrier type in the DList API with a
'groupBarrier' method. Use this method in the "join" and "cogroup"
implementations to ensure correct Grouping option semantics with respect
to Hadoop.