Remove DelayedInit from ScoobiApp.
Modify ScoobiApp to not mixin DelayedInit. Instead, a client must
implement the 'run()' method if it mixes-in ScoobiApp.

Previously, when using ScoobiApp, if the function passed to a DList
method closed over a value, the value will not be initialised at the
time when the function is run. This is an issue with the use of
DelayedInit and hence the reason for dropping it.