fixed the simple groupByKey.groupBarrier.groupByKey example
When persisting traversables, we persist 1000 elements by 1000 elements.
The bug in the previous code was partially related to iterating a first time to get the size of the iterator,
then to iterate a second time to access the elements (when the iterator is actually consumed).

The code has been refactored so that one implementation covers Traversable, Maps and Arrays