Fixed issue #113: taking the minimum and the maximum of a list
The issue was in the Persister class were the relationship between DComp nodes and their optimised version was not properly tracked.
More specifically, a DComp nodes were associated to its sinks and sinks to their optimised nodes. However in the case of materialised values
there were no sinks so when looking for the result for a particular DComp, we ended up getting the result for another one.

This whole issue indicates that we should find a better way to track the DComp <-> optimised nodes relationship.